Modularization
Broken down the code into three files where the functions are related in some way. (Untested)
